requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.9
requests==2.32.3